Added some stuff
Got a little better understanding of servlet, modified a few things.
Both Login and Logout are working effectively  for Admin, Login for the user.
Added a script for the cart, still need to fix things about the style and finish up fixing register.
Need to add order eventually, and the ability to both remove an item from the cart and check its description.