updates
Added the ability for the user to check the orders, still need to maybe add a section for the products inside Order_Details somehow because of toString printing addedToCart.
Weird stuff always happens if querryis fails resulting in stuff inside the session  getting erased, probably due to the fact that the Servlet interact with those values, can't really do anything about it.