Fixs
Fixed login kenw what the issue was, but was too lazy to fix it, gave up in the end and fixed it, everithing works fine now.
Admin can now remove item, the user can still see the cost and the removed item id is visualized for debugging purposes, Register servlet was fixed as well : needed to set the session attribute 'user' to null in a very long time fixed it today(finally). Why am i so lazy to fix things i could fix in like 5 seconds i dunno. :ant: